backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
openai-whisper==20231117
transformers==4.44.0
sentencepiece==0.2.0
torch==2.3.0
elevenlabs==1.5.0
pydub==0.25.1
soundfile==0.12.1
numpy==1.26.4
websockets==12.0
python-dotenv==1.0.1
aiofiles==24.1.0
gtts==2.5.1
